Rewrote interval lists to support future work in interval aggregation
